comments/views.py
from rest_framework.viewsets import ModelViewSet
from rest_framework.permissions import IsAuthenticatedOrReadOnly


from core.permissions import IsCreatorOrReadOnly

from .models import Comment
from .serializers import CommentSerializer


class CommentViewSet(ModelViewSet):
    queryset = Comment.objects.all()
    serializer_class = CommentSerializer
    filterset_fields = ['post_id']
    permission_classes = {
        IsAuthenticatedOrReadOnly,
        IsCreatorOrReadOnly
    }

    def perform_create(self, serializer):
        serializer.save(creator=self.request.user)
