requirements.txt
Django==2.2.3
django-environ==0.4.5
django-extensions==2.1.9
django-filter==2.2.0
djangorestframework==3.10.0
djangorestframework-stubs==0.4.2
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
